requirements.txt
# Core dependencies for the ReAct Agent project
python-dotenv>=1.0.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langchain-text-splitters>=0.1.0
langchain-chroma>=0.1.0
langgraph>=0.1.0
supabase>=2.0.0

# Date parsing
python-dateutil>=2.8.0

# FAQ System dependencies
PyPDF2>=3.0.0
reportlab>=4.0.0
chromadb>=0.4.0
openai>=1.0.0

# Optional dependencies for development
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0

# LiveKit Agents dependencies
livekit-agents[openai, silero, deepgram, cartesia]==1.2.2
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1
httpx==0.28.1
ipython==8.13.2
pydantic>=2.0.0
PyYAML>=6.0.0

# Local TTS dependencies
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
datasets>=2.0.0
librosa>=0.10.0
soundfile>=0.12.0
speechbrain>=0.5.0
pyttsx3>=2.90  # Fallback TTS

# GUI dependencies
PySide6>=6.5.0
qasync>=0.23.0


